+++ database/seeders/ProvinceSeeder.php
@@ -11,6 +11,9 @@ class ProvinceSeeder extends Seeder
 {
     public function run()
     {
+        if (! \Schema::hasTable((new Province)->getTable())) {
+            throw new \Exception('Province table does not exist.');
+        }
         Province::query()->delete();
         if (File::exists(base_path('/vendor/soap/thai-addresses/database/seeders/data/provinces.json'))) {
             $json = File::get(base_path('/vendor/soap/thai-addresses/database/seeders/data/provinces.json'));
